Move shared kernelpkg test cases to tests/support/
