+++ tests/support/kernelpkg.py
@@ -3,7 +3,7 @@
     :synopsis: Base class for kernelpkg modules
     :platform: Linux
     :maturity: develop
-    versionadded:: 2018.3.0
+    .. versionadded:: 2018.3.0
 '''
 # pylint: disable=invalid-name,no-member
